sokoconnect/requirements.txt
asgiref==3.8.1
django==5.2.3
djangorestframework==3.16.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing-extensions==4.14.0
